Create Horizon-ConnectionServers-Upgrade.ps1
I use this script to automatically upgrade the Connection Servers in a pod, it will talk to the primary-cs and get a list of CS that are installed (primary and replicas) then go through the list and upgrade one at a time.  Does require to auto login in Windows due to Kerberos limitations with running remotely.